backend/pyproject.toml
[project]
name = "network-synapse-quattro"
version = "0.1.0"
description = "Network Synapse Quattro backend — Infrahub SoT interaction, config generation, schema management"
requires-python = ">=3.12"
license = "Apache-2.0"
dependencies = [
    "infrahub-sdk[ctl]>=1.0.0",
    "jinja2>=3.1",
    "pydantic>=2.5",
    "httpx>=0.25",
    "pyyaml>=6.0",
    "pygnmi>=0.8",
    "grpcio>=1.60",
    "nornir>=3.4,<4.0",
    "nornir-utils>=0.2",
    "scrapli>=2024.1",
    "scrapli-community>=2024.1",
    "netmiko>=4.3",
    "napalm>=4.1",
    "python-json-logger>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["network_synapse"]
